+++ lib/src/manage_sell_offer_operation.dart
@@ -2,6 +2,8 @@
 // Use of this source code is governed by a license that can be
 // found in the LICENSE file.
 
+import 'package:stellar_flutter_sdk/src/muxed_account.dart';
+
 import 'operation.dart';
 import 'assets.dart';
 import 'price.dart';
@@ -84,7 +86,7 @@ class ManageSellOfferOperationBuilder {
   String _amount;
   String _price;
   int _offerId = 0;
-  String _mSourceAccount;
+  MuxedAccount _mSourceAccount;
 
   /// Creates a new ManageSellOfferOperation builder. If you want to update existing offer use
   ManageSellOfferOperationBuilder(
@@ -101,8 +103,16 @@ class ManageSellOfferOperationBuilder {
     return this;
   }
 
-  ///Sets the source account for this operation.
+  /// Sets the source account for this operation.
   ManageSellOfferOperationBuilder setSourceAccount(String sourceAccount) {
+    checkNotNull(sourceAccount, "sourceAccount cannot be null");
+    _mSourceAccount = MuxedAccount(sourceAccount, null);
+    return this;
+  }
+
+  /// Sets the muxed source account for this operation.
+  ManageSellOfferOperationBuilder setMuxedSourceAccount(
+      MuxedAccount sourceAccount) {
     _mSourceAccount =
         checkNotNull(sourceAccount, "sourceAccount cannot be null");
     return this;