+++ lib/src/path_payment_strict_send_operation.dart
@@ -2,6 +2,8 @@
 // Use of this source code is governed by a license that can be
 // found in the LICENSE file.
 
+import 'package:stellar_flutter_sdk/src/muxed_account.dart';
+
 import 'operation.dart';
 import 'assets.dart';
 import 'key_pair.dart';
@@ -16,13 +18,18 @@ import 'xdr/xdr_type.dart';
 class PathPaymentStrictSendOperation extends Operation {
   Asset _sendAsset;
   String _sendAmount;
-  String _destination;
+  MuxedAccount _destination;
   Asset _destAsset;
   String _destMin;
   List<Asset> _path;
 
-  PathPaymentStrictSendOperation(Asset sendAsset, String sendAmount,
-      String destination, Asset destAsset, String destMin, List<Asset> path) {
+  PathPaymentStrictSendOperation(
+      Asset sendAsset,
+      String sendAmount,
+      MuxedAccount destination,
+      Asset destAsset,
+      String destMin,
+      List<Asset> path) {
     this._sendAsset = checkNotNull(sendAsset, "sendAsset cannot be null");
     this._sendAmount = checkNotNull(sendAmount, "sendAmount cannot be null");
     this._destination = checkNotNull(destination, "destination cannot be null");
@@ -44,7 +51,7 @@ class PathPaymentStrictSendOperation extends Operation {
   String get sendAmount => _sendAmount;
 
   /// Account that receives the payment.
-  String get destination => _destination;
+  MuxedAccount get destination => _destination;
 
   /// The asset the destination account receives.
   Asset get destAsset => _destAsset;
@@ -66,7 +73,7 @@ class PathPaymentStrictSendOperation extends Operation {
     sendMax.int64 = Operation.toXdrAmount(this.sendAmount);
     op.sendMax = sendMax;
     // destination
-    op.destination = KeyPair.fromAccountId(this._destination).xdrMuxedAccount;
+    op.destination = this._destination.toXdr();
     // destAsset
     op.destAsset = destAsset.toXdr();
     // destAmount
@@ -106,23 +113,39 @@ class PathPaymentStrictSendOperation extends Operation {
 class PathPaymentStrictSendOperationBuilder {
   Asset _sendAsset;
   String _sendAmount;
-  String _destination;
+  MuxedAccount _destination;
   Asset _destAsset;
   String _destMin;
   List<Asset> _path;
-  String _mSourceAccount;
+  MuxedAccount _mSourceAccount;
 
   /// Creates a PathPaymentStrictSendOperation builder.
   PathPaymentStrictSendOperationBuilder(Asset sendAsset, String sendAmount,
       String destination, Asset destAsset, String destMin) {
     this._sendAsset = checkNotNull(sendAsset, "sendAsset cannot be null");
     this._sendAmount = checkNotNull(sendAmount, "sendAmount cannot be null");
-    this._destination = checkNotNull(destination, "destination cannot be null");
+    checkNotNull(destination, "destination cannot be null");
+    this._destination = MuxedAccount(destination, null);
     this._destAsset = checkNotNull(destAsset, "destAsset cannot be null");
     this._destMin = checkNotNull(destMin, "destMin cannot be null");
   }
 
-  ///Sets path for this operation
+  /// Creates a PathPaymentStrictSendOperation builder for a MuxedAccount as a destination.
+  PathPaymentStrictSendOperationBuilder.forMuxedDestinationAccount(
+      Asset sendAsset,
+      String sendAmount,
+      MuxedAccount destination,
+      Asset destAsset,
+      String destMin) {
+    this._sendAsset = checkNotNull(sendAsset, "sendAsset cannot be null");
+    this._sendAmount = checkNotNull(sendAmount, "sendAmount cannot be null");
+    checkNotNull(destination, "destination cannot be null");
+    this._destination = destination;
+    this._destAsset = checkNotNull(destAsset, "destAsset cannot be null");
+    this._destMin = checkNotNull(destMin, "destMin cannot be null");
+  }
+
+  /// Sets path for this operation
   PathPaymentStrictSendOperationBuilder setPath(List<Asset> path) {
     checkNotNull(path, "path cannot be null");
     checkArgument(
@@ -131,10 +154,18 @@ class PathPaymentStrictSendOperationBuilder {
     return this;
   }
 
-  ///Sets the source account for this operation.
+  /// Sets the source account for this operation.
   PathPaymentStrictSendOperationBuilder setSourceAccount(String sourceAccount) {
-    _mSourceAccount =
-        checkNotNull(sourceAccount, "sourceAccount cannot be null");
+    checkNotNull(sourceAccount, "sourceAccount cannot be null");
+    _mSourceAccount = MuxedAccount(sourceAccount, null);
+    return this;
+  }
+
+  /// Sets the muxed source account for this operation.
+  PathPaymentStrictSendOperationBuilder setMuxedSourceAccount(
+      MuxedAccount sourceAccount) {
+    checkNotNull(sourceAccount, "sourceAccount cannot be null");
+    _mSourceAccount = sourceAccount;
     return this;
   }