+++ lib/stellar_flutter_sdk.dart
@@ -72,6 +72,7 @@ export 'src/responses/effects/misc_effects_responses.dart';
 
 // Stellar
 export 'src/account.dart';
+export 'src/muxed_account.dart';
 export 'src/account_flag.dart';
 export 'src/assets.dart';
 export 'src/asset_type_native.dart';